Add soak test for fault-tolerant S2S join [HZ-1991] (#176)
This test creates a SQL job that performs an equi-join on 2 streaming
sources (views of the same Kafka source) and sinks the result to a Kafka
topic. Events are inserted into the source topic at regular intervals,
and verification is performed via `KafkaSinkVerifier` to confirm that
duplicate events weren't generated by the S2S-join job
(processingGuarantee=exactlyOnce).